+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Handle location processing on WebTools import [#90](https://github.com/rokwire/gateway-building-block/issues/90)
 
 [2.9.0] - 2024-05-07
 ### Changed